fix analytics placeholder text shown when no bills are added
