+++ src/contexts/useBillyContext.ts
@@ -7,11 +7,13 @@ import BillService from '../services/BillService';
 import Cache, {STORAGE_KEYS} from '../services/Cache';
 import UserService from '../services/UserService';
 import {Context} from './BillyContext';
+import {getLastBillDate} from '../helpers/AnalyticsFns';
 
 export const useBilly = (): Context => {
   const [bills, setBills] = useState<Bill[]>([]);
   const [upcomingBills, setUpcomingBills] = useState<Bill[]>([]);
   const [missedBills, setMissedBills] = useState<Bill[]>([]);
+  const [latestBillDate, setLatestBillDate] = useState<string | null>(null);
   const [lastSyncDate, setLastSyncDate] = useState<string>('');
   const [reminders, setReminders] = useState<Record<string, number>>({});
   const [user, setUser] = useState<User | null>(null);
@@ -50,6 +52,7 @@ export const useBilly = (): Context => {
     getBillIdToNumRemindersMap(bills).then(retrievedReminders => {
       setReminders(retrievedReminders);
     });
+    setLatestBillDate(getLastBillDate(bills).format('DD MMM YYYY'));
   }, [bills]);
 
   const setCurrentBills = useCallback((items: Bill[]) => {
@@ -60,6 +63,7 @@ export const useBilly = (): Context => {
     bills,
     missedBills,
     upcomingBills,
+    latestBillDate,
     reminders,
     lastSyncDate,
     user,
